remove packages and migrate [realm]/[auctionHouse] back to page router
