+++ package.json
@@ -13,11 +13,9 @@
 		"mongodb": "^4.1.3",
 		"mongoose": "^6.5.2",
 		"next": "^14.0.4",
-		"patch-package": "^8.0.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-intersection-observer": "^9.5.3",
-		"sharp": "^0.32.3",
 		"typewriter-effect": "^2.21.0"
 	},
 	"devDependencies": {
