+++ utils/clients/blizzard/client.js
@@ -1,3 +1,4 @@
+import { revalidate } from '../../../app/page';
 import getAccessToken from '../../db/getAccessToken';
 import updateAccessToken from '../../db/updateAccessToken';
